adding new chnages

+++ views/Tags.php
@@ -12,7 +12,7 @@
                 <ul class="nav nav-pills flex-column mb-auto">
 
                     <li>
-                        <a href="#" class="nav-link link-body-emphasis ">
+                        <a href="/profile" class="nav-link link-body-emphasis ">
                             <i class="bi bi-speedometer2"></i>
                             Dashboard
                         </a>
@@ -27,13 +27,13 @@
                         </li>
                     <?php }?>
                     <li>
-                        <a href="#" class="nav-link link-body-emphasis active text-white">
+                        <a href="/addTag" class="nav-link link-body-emphasis active text-white">
                             <i class="bi bi-tags"></i>
                             Tags
                         </a>
                     </li>
                     <li>
-                        <a href="#" class="nav-link link-body-emphasis">
+                        <a href="/categories" class="nav-link link-body-emphasis">
                             <i class="bi bi-bookmarks"></i>
                             Categories
                         </a>
@@ -48,32 +48,47 @@
         <div class="col-sm-9 col-sm-offset-3 col-md-10 col-md-offset-2 main">
             <h1 class="page-header">Dashboard</h1>
             <h1 class="page-header text-white">Dashboard</h1>
+            <?php if($_SESSION['role'] == 'Admin') {?>
             <h2 class="my-5"> Add Tag</h2>
+                <form method="POST">
+                    <label for="exampleDataList" class="form-label">Insert Your Tag</label>
+                    <input class="form-control editTagName" list="datalistOptions" name="tag" id="exampleDataList" placeholder="Add Tag">
+                    <input type="hidden" class="form-control editTag" list="datalistOptions" name="edit" id="exampleDataListEdit" placeholder=" Tag">
+                    <button class="btn btn-primary my-3" type="submit" name="submit" value="add">Add Tag</button>
+                    <button class="btn btn-success my-3" type="submit" name="submit" value="edit">edit Tag</button>
 
-            <form method="POST">
-                <label for="exampleDataList" class="form-label">Insert Your Tag</label>
-                <input class="form-control" list="datalistOptions" name="tag" id="exampleDataList" placeholder="Add Tag">
-                <button class="btn btn-primary my-3" type="submit" name="submit">Add Tag</button>
-            </form>
+                </form>
+            <?php }?>
+            <?php if($_SESSION['role'] == 'Reader'){?>
+                <form method="POST">
+                    <label for="exampleDataList" class="form-label">Search Tag</label>
+                    <input class="form-control" list="datalistOptions" name="tag" id="exampleDataList" placeholder="Type to search">
+                </form>
+            <?php }?>
 
-            <h2 class="sub-header">Users</h2>
+            <h2 class="sub-header my-5">Tags</h2>
             <div class="table-responsive">
                 <table class="table table-striped">
                     <thead>
                     <tr>
                         <th>#</th>
                         <th>Tag Name</th>
-                        <th>Action</th>
+                        <?php if($_SESSION['role'] == 'Admin') {?>
+                            <th>Action</th>
+                        <?php }?>
 
                     </tr>
                     </thead>
                     <tbody>
                     <?php foreach ($tags as $tag): ?>
 
-                        <tr>
+                        <tr style="cursor: pointer">
                             <td><?= $tag['TagID']?></td>
-                            <td><?= $tag['TagName']?></td>
-                            <td><a href="/deleteTag?id=<?=$tag['TagID']?>" class="btn btn-danger"><i class="bi bi-trash3"></i></a></td>
+                            <td class="tagName"><?= $tag['TagName']?></td>
+                            <?php if($_SESSION['role'] == 'Admin') {?>
+                                <td><a href="/deleteTag?id=<?=$tag['TagID']?>" class="btn btn-danger" ><i class="bi bi-trash3"></i></a></td>
+                            <?php }?>
+
                         </tr>
                     <?php endforeach; ?>
 
@@ -87,3 +102,20 @@
     </div>
 </div>
 
+<script>
+
+    let editTag = document.querySelector('.editTag');
+    let editTagName = document.querySelector('.editTagName');
+    let tableTr = document.querySelectorAll('tbody tr');
+
+    tableTr.forEach((tr) =>{
+        tr.addEventListener('click', function ()
+        {
+            console.log(tr);
+            editTag.value = tr.querySelector('td').textContent;
+            editTagName.value = tr.querySelector('.tagName').textContent;
+        })
+    })
+
+
+</script>
